+++ lib/flutter_quill_to_pdf.dart
@@ -11,7 +11,7 @@ export 'package:flutter_quill_to_pdf/core/extensions/pdf_extension.dart';
 export 'package:flutter_quill_to_pdf/core/extensions/string_extension.dart';
 export 'package:flutter_quill_to_pdf/packages/html2md/lib/html2md.dart';
 export 'package:flutter_quill_to_pdf/packages/vsc_quill_delta_to_html/vsc_quill_delta_to_html.dart';
-export 'package:flutter_quill_to_pdf/domain/entities/custom_converter.dart';
+export 'package:flutter_quill_to_pdf/converter/configurator/converter_option/custom_converter.dart';
 export 'package:flutter_quill_to_pdf/utils/converters_utils.dart';
 export 'package:flutter_quill_to_pdf/utils/makdown_rules_custom.dart';
 export 'package:flutter_quill_to_pdf/utils/typedefs.dart';